requirements.txt
langfuse>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.5.0
boto3>=1.28.0
langchain>=0.0.267
langchain-openai>=0.0.5 